refactor: FormatState -> State
